Package "general" renamed to "math"
